feat(Compliance): split fec config file based on env
