+++ fec.config.js
@@ -1,114 +1,4 @@
-const { resolve } = require('path');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const alias = require('./config/aliases');
-const packageJson = require('./package.json');
-const { localRoutesFor } = require('./config/helpers');
-const { sentryWebpackPlugin } = require('@sentry/webpack-plugin');
-
-const bundle = 'insights';
-const appName = packageJson[bundle].appname;
-const isIOP = process.env.IOP === 'true';
-
-const iopAliases = isIOP
-  ? {
-      '@redhat-cloud-services/frontend-components/InsightsLink': resolve(
-        __dirname,
-        './src/iop/ComplianceLink.js',
-      ),
-      '@redhat-cloud-services/frontend-components-utilities/useInsightsNavigate':
-        resolve(__dirname, './src/iop/useComplianceNavigate.js'),
-      react: resolve(__dirname, 'node_modules/react'),
-      'react-dom': resolve(__dirname, 'node_modules/react-dom'),
-      'react/jsx-runtime': resolve(
-        __dirname,
-        'node_modules/react/jsx-runtime',
-      ),
-    }
-  : {};
-
-// TODO Move to fec webpack config similar to LOCAL_APPS
-const routes = {
-  ...(process.env.LOCAL_APIS && process.env.LOCAL_APIS !== ''
-    ? localRoutesFor('/api', process.env.LOCAL_APIS)
-    : {}),
-};
-
-module.exports = {
-  appName,
-  appUrl: isIOP ? '/' : `/${bundle}/${appName}`,
-  definePlugin: isIOP ? { 'process.env.IOP': JSON.stringify('true') } : {},
-  useProxy: process.env.PROXY === 'true',
-  ...(isIOP
-    ? {
-        deployment: 'assets/apps',
-        standalone: true,
-        appEntry: resolve(__dirname, './src/iop/entry.js'),
-      }
-    : {}),
-  devtool: 'hidden-source-map',
-  plugins: [
-    ...(isIOP
-      ? [
-          new HtmlWebpackPlugin({
-            template: resolve(__dirname, './src/iop/index.html'),
-            inject: true,
-            chunks: ['App'],
-          }),
-        ]
-      : []),
-    ...(process.env.ENABLE_SENTRY
-      ? [
-          sentryWebpackPlugin({
-            ...(process.env.SENTRY_AUTH_TOKEN && {
-              authToken: process.env.SENTRY_AUTH_TOKEN,
-            }),
-            org: 'red-hat-it',
-            project: 'compliance-rhel',
-            moduleMetadata: ({ release }) => ({
-              dsn: `https://6410c806f0ac7b638105bb4e15eb3399@o490301.ingest.us.sentry.io/4508083145408512`,
-              org: 'red-hat-it',
-              project: 'compliance-rhel',
-              release,
-            }),
-          }),
-        ]
-      : []),
-  ],
-  moduleFederation: isIOP
-    ? {
-        exposes: {},
-        bundleChromeShared: true,
-      }
-    : {
-        shared: [
-          {
-            'react-router-dom': {
-              singleton: true,
-              import: false,
-              version: packageJson.dependencies['react-router-dom'],
-              requiredVersion: '>=6.0.0 <7.0.0',
-            },
-          },
-        ],
-        exposes: {
-          './RootApp': resolve(
-            __dirname,
-            `/src/${process.env.NODE_ENV !== 'production' ? 'Dev' : ''}AppEntry`,
-          ),
-          './SystemDetail': resolve(__dirname, '/src/Modules/ComplianceDetails'),
-          './ReportPDFBuild': resolve(
-            __dirname,
-            '/src/SmartComponents/ExportPDF/ReportPDFBuild',
-          ),
-        },
-      },
-  resolve: {
-    alias: {
-      ...alias,
-      ...iopAliases,
-    },
-  },
-  routes,
-  _unstableSpdy: true,
-  frontendCRDPath: 'deployment_template.yaml',
-};
+module.exports =
+  process.env.IOP === 'true'
+    ? require('./config/fec.iop')
+    : require('./config/fec.hcc');
